+++ src/admin/voluntary/dto/update-voluntary.dto.ts
@@ -1,5 +1,5 @@
 import { PartialType } from '@nestjs/mapped-types';
-import { CreateVoluntarioDto } from './create-voluntario.dto';
+import { CreateVoluntarioDto } from './create-voluntary.dto';
 
 export class UpdateVoluntarioDto extends PartialType(CreateVoluntarioDto) {
   id: number;
