+++ src/client/responsible/dto/create-responsible.dto.ts
@@ -2,7 +2,6 @@ import { IsString, IsNumber, IsDate, IsBoolean } from "class-validator";
 import { ApiProperty } from "@nestjs/swagger";
 
 export class CreateResponsibleDto {
-    
     @ApiProperty()
     @IsString()
     name: string;
